# Improve search performance when there are many search results