--- a/assets/js/search.js
+++ b/assets/js/search.js
@@ -69,17 +69,26 @@ const renderResults = (results) => {
     self.findIndex(e => e.id === element.id) === index)
 
   let instance = new Mark(document.querySelector('#searchResults'))
-  let elements = arr.map((result) => {
-    const matchPos = result.doc.body.indexOf(query)
-    const bodyStartPos = matchPos - BODY_LENGTH / 2 > 0 ? matchPos - BODY_LENGTH / 2 : 0
-    const body = result.doc.body.substr(bodyStartPos, BODY_LENGTH)
+  let fragment = document.createDocumentFragment();
+  arr.forEach((result) => {
+    let resultPage = document.createElement('div')
+    resultPage.className = 'searchResultPage'
 
-    return `<div class="searchResultPage">
-        <a class="searchResultTitle" href="${result.doc.href}">${result.doc.title}</a>
-        <div class="searchResultBody">${body}</div>
-      </div>`
+    let resultTitle = document.createElement('a')
+    resultTitle.className = 'searchResultTitle'
+    resultTitle.href = result.doc.href
+    resultTitle.innerHTML = result.doc.title
+    resultPage.append(resultTitle)
+
+    let resultBody = document.createElement('div')
+    resultBody.className = 'searchResultBody'
+    let matchPos = result.doc.body.indexOf(query)
+    let bodyStartPos = matchPos - BODY_LENGTH / 2 > 0 ? matchPos - BODY_LENGTH / 2 : 0
+    resultBody.innerHTML = result.doc.body.substr(bodyStartPos, BODY_LENGTH)
+    resultPage.append(resultBody)
+    fragment.append(resultPage)
   })
-  searchResults.innerHTML = elements.join('');
+  searchResults.append(fragment);
   instance.mark(query)
 }