# Issue

When there are many search results, searching is very slow.

In the case of the following, searching freezes the page for nearly 20 seconds.
* **Size of `search-data.json`**: 1.1MB
* **Number of blog entries**: 128
* **Number of search results**: 98

https://github.com/matsuyoshi30/harbor/assets/1453749/86281056-82ae-4855-9b35-354cac038a44

# Fixes

Since the issue is caused by the way to append html element, this PR fix it. With this fix, performance in the same situation is improved.

https://github.com/matsuyoshi30/harbor/assets/1453749/7a7e03a6-4198-4efa-9712-7bce4a8f779a