@matsuyoshi30 2023-07-04
First of all, thank you for your contribution!

I believe it's better to use the DOM API, like `createElement`, for generating DOM elements rather than generating them directly from a string. In this case, the issue likely arises from creating elements with createElement and appending them within each loop.

When adding a large number of elements to the DOM at once, we can reduce the number of reflows and repaints by first adding them to a `DocumentFragment`, rather than adding each element individually. Afterward, the entire DocumentFragment can be added to the DOM all at once.
Could you check what the result would be when using [createDocumentFragment](https://developer.mozilla.org/en-US/docs/Web/API/Document/createDocumentFragment)?